refactor(): cookie-util 클래스명 변경 및 쿠키 max-age 변경
- 쿠키의 age는 무조건 그날 하루로 설정
- ViewHistoryManager로 이름 변경
